+++ src/extension.ts
@@ -1,6 +1,7 @@
 import * as log from './log';
 import * as vscode from 'vscode';
-import { TarFsProvider, ZipFsProvider } from './fs-providers';
+import { TarFsProvider, WorkspaceFolderStore, ZipFsProvider } from './fs-providers';
+import { cleanTempDir } from './util';
 
 export function activate(context: vscode.ExtensionContext) {
     log.init(context);
@@ -8,4 +9,7 @@ export function activate(context: vscode.ExtensionContext) {
     ZipFsProvider.register(context);
 }
 
-export function deactivate() {}
+export function deactivate() {
+    cleanTempDir();
+    WorkspaceFolderStore.getInstance().clean();
+}