+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to the "archive-inspector" extension will be documented in t
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [Unreleased]
+
+- Support nested archives via extraction to temporary file.
+
 ## [0.2.0]
 
 - Added `.zip` support.